% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/commandinput.R
\name{commandInput}
\alias{commandInput}
\title{Command input}
\usage{
commandInput(inputId, label, value, ..., autoclear = TRUE)
}
\arguments{
\item{inputId}{Input variable to assign the control's value to.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{value}{Initial value.}

\item{...}{Additional attributes to be included on the \code{<input>} tag.}

\item{autoclear}{If \code{TRUE}, the contents of the text box are cleared
  after the value is sent.}
}
\description{
Creates a text box that only sends its value when the user hits Enter.
}

