% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/tableinput.R
\name{matrixInput}
\alias{matrixInput}
\title{Matrix input}
\usage{
matrixInput(inputId, label, data, datatype = "character")
}
\arguments{
\item{inputId}{Input variable to assign the control's value to.}

\item{label}{Display label for the control.}

\item{data}{The initial values to use for the matrix.}

\item{datatype}{The type of data the input dataframe will except, defaults now to \code{character}.}
}
\description{
Creates an adjustable-length matrix input.
}

