% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/durationInput.R
\name{durationInput}
\alias{durationInput}
\title{Duration Input}
\usage{
durationInput(inputId, label, value = list(hour = 0, minute = 0, second = 0),
  step = NULL, width = "50px")
}
\arguments{
\item{inputId}{Input variable to assign the control's value to.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{value}{Initial value as number of seconds or a list of values for hour, minute, second}

\item{width}{CSS width property}
}
\description{
Creates a set of linked numeric inputs that represent hour, minute and seconds
Returns duration in seconds.
}

