\name{matrixInput}
\alias{matrixInput}
\title{Matrix input}
\usage{
  matrixInput(inputId, label, data)
}
\arguments{
  \item{inputId}{Input variable to assign the control's
  value to.}

  \item{label}{Display label for the control.}

  \item{data}{The initial values to use for the matrix.}
}
\description{
  Creates an adjustable-length matrix input.
}

